# Rewrite Uri.EscapeString