--- a/src/System.Private.Uri/src/System/UriHelper.cs
+++ b/src/System.Private.Uri/src/System/UriHelper.cs
@@ -744,6 +744,8 @@ namespace System
             }
             return false;
         }
+        
+        // "Reserved" and "Unreserved" characters are based on RFC 3986.
 
         internal static ReadOnlySpan<bool> UnreservedReservedTable => new bool[0x80]
         {