Several public methods (Uri.EscapeDataString, Uri.EscapeUriString) and a bunch of internal call sites rely on the internal EscapeString helper.  This helper has several issues with it:
- It uses unsafe code.
- It unnecessarily requires and copies through a char[] to get to a string when a string is the required result.
- It has a lot of complexity around the handling of unicode.

This PR rewrites it to utilize Span, Rune, and other newer features in a way that enables it to be both safe and efficient.  Most inputs ends up being faster, and for very long inputs, it's much, much faster.  The use of ValueStringBuilder also results in less memory allocation, in some cases significantly.

The use of Rune also fixes two arguable bugs in the existing implementation around invalid Unicode sequences, which is why a couple tests were tweaked:
- Some but not all invalid unicode patterns result in replacement characters being used: a few invalid sequences (e.g. just a high surrogate) result in an exception.  We should be standardized on using replacement characters for all such invalid sequences.
- Some patterns with invalid unicode patterns actually result in unnecessary encoding, e.g. `Uri.EscapeDataString("\uD800\uD800a")` results in `a` being encoded.

|     Method | Tool | Length |                 Kind |                Mean | Ratio |    Allocated |
|----------- |---------- |------- |--------------------- |--------------------:|------:|-------------:|
| EscapeData | new       |     10 |           Unreserved |            19.92 ns |  0.23 |            - |
| EscapeData | old       |     10 |           Unreserved |            84.86 ns |  1.00 |            - |
|            |           |        |                      |                     |       |              |
|  EscapeUri | new       |     10 |           Unreserved |            17.00 ns |  0.17 |            - |
|  EscapeUri | old       |     10 |           Unreserved |           103.09 ns |  1.00 |            - |
|            |           |        |                      |                     |       |              |
| EscapeData | new       |     10 |              Unicode |           316.53 ns |  0.96 |        264 B |
| EscapeData | old       |     10 |              Unicode |           330.18 ns |  1.00 |       1248 B |
|            |           |        |                      |                     |       |              |
|  EscapeUri | new       |     10 |              Unicode |           313.91 ns |  0.95 |        264 B |
|  EscapeUri | old       |     10 |              Unicode |           328.76 ns |  1.00 |       1248 B |
|            |           |        |                      |                     |       |              |
| EscapeData | new       |     10 | OneRe(...)erved [25] |           154.93 ns |  1.04 |         48 B |
| EscapeData | old       |     10 | OneRe(...)erved [25] |           149.36 ns |  1.00 |        312 B |
|            |           |        |                      |                     |       |              |
|  EscapeUri | new       |     10 | OneRe(...)erved [25] |           152.34 ns |  0.85 |         48 B |
|  EscapeUri | old       |     10 | OneRe(...)erved [25] |           179.05 ns |  1.00 |        312 B |
|            |           |        |                      |                     |       |              |
| EscapeData | new       |     10 |          Alternating |           192.87 ns |  0.73 |        120 B |
| EscapeData | old       |     10 |          Alternating |           262.74 ns |  1.00 |        392 B |
|            |           |        |                      |                     |       |              |
|  EscapeUri | new       |     10 |          Alternating |           189.67 ns |  0.65 |        120 B |
|  EscapeUri | old       |     10 |          Alternating |           289.09 ns |  1.00 |        392 B |
|            |           |        |                      |                     |       |              |
| EscapeData | new       |    100 |           Unreserved |            83.99 ns |  0.13 |            - |
| EscapeData | old       |    100 |           Unreserved |           660.66 ns |  1.00 |            - |
|            |           |        |                      |                     |       |              |
|  EscapeUri | new       |    100 |           Unreserved |            83.69 ns |  0.09 |            - |
|  EscapeUri | old       |    100 |           Unreserved |           943.46 ns |  1.00 |            - |
|            |           |        |                      |                     |       |              |
| EscapeData | new       |    100 |              Unicode |         3,015.49 ns |  0.96 |       2424 B |
| EscapeData | old       |    100 |              Unicode |         3,150.40 ns |  1.00 |      12144 B |
|            |           |        |                      |                     |       |              |
|  EscapeUri | new       |    100 |              Unicode |         3,004.75 ns |  0.90 |       2424 B |
|  EscapeUri | old       |    100 |              Unicode |         3,334.80 ns |  1.00 |      12144 B |
|            |           |        |                      |                     |       |              |
| EscapeData | new       |    100 | OneRe(...)erved [25] |           997.48 ns |  0.96 |        232 B |
| EscapeData | old       |    100 | OneRe(...)erved [25] |         1,034.85 ns |  1.00 |        496 B |
|            |           |        |                      |                     |       |              |
|  EscapeUri | new       |    100 | OneRe(...)erved [25] |         1,009.19 ns |  0.89 |        232 B |
|  EscapeUri | old       |    100 | OneRe(...)erved [25] |         1,126.74 ns |  1.00 |        496 B |
|            |           |        |                      |                     |       |              |
| EscapeData | new       |    100 |          Alternating |         1,705.95 ns |  0.72 |       1080 B |
| EscapeData | old       |    100 |          Alternating |         2,361.35 ns |  1.00 |       2536 B |
|            |           |        |                      |                     |       |              |
|  EscapeUri | new       |    100 |          Alternating |         1,690.50 ns |  0.63 |       1080 B |
|  EscapeUri | old       |    100 |          Alternating |         2,691.28 ns |  1.00 |       2536 B |
|            |           |        |                      |                     |       |              |
| EscapeData | new       | 100000 |           Unreserved |        66,052.21 ns |  0.09 |            - |
| EscapeData | old       | 100000 |           Unreserved |       710,216.92 ns |  1.00 |            - |
|            |           |        |                      |                     |       |              |
|  EscapeUri | new       | 100000 |           Unreserved |        66,393.07 ns |  0.08 |            - |
|  EscapeUri | old       | 100000 |           Unreserved |       874,426.80 ns |  1.00 |          1 B |
|            |           |        |                      |                     |       |              |
| EscapeData | new       | 100000 |              Unicode |     4,295,295.08 ns | 0.003 |    6594352 B |
| EscapeData | old       | 100000 |              Unicode | 1,490,528,971.72 ns | 1.000 | 6006095568 B |
|            |           |        |                      |                     |       |              |
|  EscapeUri | new       | 100000 |              Unicode |     4,214,907.36 ns | 0.003 |    6594359 B |
|  EscapeUri | old       | 100000 |              Unicode | 1,517,378,125.77 ns | 1.000 | 6006096360 B |
|            |           |        |                      |                     |       |              |
| EscapeData | new       | 100000 | OneRe(...)erved [25] |       939,019.21 ns |  0.94 |     200033 B |
| EscapeData | old       | 100000 | OneRe(...)erved [25] |       984,792.58 ns |  1.00 |     400328 B |
|            |           |        |                      |                     |       |              |
|  EscapeUri | new       | 100000 | OneRe(...)erved [25] |       966,883.99 ns |  0.89 |     200040 B |
|  EscapeUri | old       | 100000 | OneRe(...)erved [25] |     1,082,288.94 ns |  1.00 |     400331 B |
|            |           |        |                      |                     |       |              |
| EscapeData | new       | 100000 |          Alternating |     1,751,153.61 ns |  0.01 |    1066679 B |
| EscapeData | old       | 100000 |          Alternating |   146,994,750.75 ns |  1.00 |  615528622 B |
|            |           |        |                      |                     |       |              |
|  EscapeUri | new       | 100000 |          Alternating |     1,735,597.85 ns |  0.01 |    1066676 B |
|  EscapeUri | old       | 100000 |          Alternating |   151,358,658.00 ns |  1.00 |  615528920 B |

```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Diagnosers;
using BenchmarkDotNet.Running;
using System;
using System.Linq;
using System.Text;

[MemoryDiagnoser]
public class Program
{
    static void Main(string[] args) => BenchmarkSwitcher.FromTypes(new[] { typeof(Program) }).Run(args);

    [Params(10, 100, 100_000)]
    public int Length { get; set; }

    [Params(InputKind.Unreserved, InputKind.Unicode, InputKind.OneReservedThenUnreserved, InputKind.Alternating)]
    public InputKind Kind { get; set; }

    public enum InputKind
    {
        Unreserved, Unicode, OneReservedThenUnreserved, Alternating
    }

    [GlobalSetup]
    public void Setup()
    {
        switch (Kind)
        {
            case InputKind.Unreserved: _input = new string('s', Length); break;
            case InputKind.Unicode: _input = string.Concat(Enumerable.Repeat("\xD83D\xDE00", Length)); break;
            case InputKind.OneReservedThenUnreserved: _input = "<" + new string('s', Length - 1); break;
            case InputKind.Alternating:
                var sb = new StringBuilder(Length);
                for (int i = 0; i < Length; i++)
                {
                    switch (i % 3)
                    {
                        case 0: sb.Append('s'); break;
                        case 1: sb.Append('<'); break;
                        default: sb.Append("\xD83D\xDE00"); break;
                    }
                }
                _input = sb.ToString();
                break;
        }
    }

    private string _input;

    [Benchmark] public string EscapeData() => Uri.EscapeDataString(_input);
    [Benchmark] public string EscapeUri() => Uri.EscapeUriString(_input);
}
```

cc: @davidsh, @GrabYourPitchforks, @alnikola 

@alnikola, I realize this conflicts with your PR.  Apologies.  But seeing your PR is what led me to want to do this.  I suggest we look at doing something similar to this PR for the "Unescape" paths as well.